+++ HISTORY.rst
@@ -2,11 +2,16 @@
 History
 =======
 
-1.2.0 (Unreleased)
+1.1.2 (Unreleased)
 ------------------
+
+* BODY-53: Return additional metadata from Image (ericof).
+* Add timeout to thumbor connection (ericof).
+
 
 1.1.1 (2016-09-28)
 ------------------
+
 * BODY-52: Quote filename for image signature (ericof).
 * BODY-55: Improve briefy.common BaseEvent (sqs event dispatcher). (rudaporto)